Added profile image display and follow/following feature

+++ InbyeolClone/app/src/main/java/com/lpennavic/inbyeolclone/model/PosterModel.kt
@@ -2,7 +2,7 @@ package com.lpennavic.inbyeolclone.model
 
 data class PosterModel(
     var description: String = "",
-    var favCnt: Int = 0,
+    var favCnt: Long = 0,
     var favs: MutableMap<String, Boolean>? = null, // 좋아요 적용 취소 체크
     var imageUrl: String = "", //
     var timestamp: String = "",
